+++ packages/client/src/components/admin-layout/Collections/index.tsx
@@ -11,8 +11,7 @@ import { PlusOutlined, DownOutlined } from '@ant-design/icons';
 import cls from 'classnames';
 import { uid, isValid, clone } from '@formily/shared';
 import { Resource } from '../../../resource';
-import { TableRowContext } from '../../../schemas/table/context';
-import { useTable } from '../../../schemas/table/hooks/useTable';
+import { TableRowContext, useTable } from '../../../schemas/table';
 import { useRequest } from 'ahooks';
 import { VisibleContext } from '../../../context';
 import { connect, FormProvider, observer, useForm } from '@formily/react';
