+++ packages/client/src/components/admin-layout/Permissions/ActionPermissionField.tsx
@@ -8,8 +8,8 @@ import { LockOutlined } from '@ant-design/icons';
 import cls from 'classnames';
 import { uid } from '@formily/shared';
 import { Resource } from '../../../resource';
-import { TableRowContext } from '../../../schemas/table/context';
-import { useTable } from '../../../schemas/table/hooks/useTable';
+import { TableRowContext } from '../../../schemas/table';
+import { useTable } from '../../../schemas/table';
 import { useRequest } from 'ahooks';
 import { VisibleContext } from '../../../context';
 import { connect, ISchema, observer, useField } from '@formily/react';
