+++ packages/client/src/schemas/calendar/DesignableBar.tsx
@@ -38,7 +38,7 @@ import { useEffect } from 'react';
 import { uid } from '@formily/shared';
 import { getSchemaPath } from '../../components/schema-renderer';
 import { useCollection, useCollectionContext } from '../../constate';
-import { useTable } from '../table/hooks/useTable';
+import { useTable } from '../table';
 import { fieldsToFilterColumns } from './';
 import { set } from 'lodash';
 import { DragHandle } from '../../components/Sortable';
