+++ packages/client/src/schemas/table/index.ts
@@ -1,0 +1,5 @@
+export * from './Table';
+export * from './context';
+
+//hooks export
+export * from './hooks/useTable';
